removing sensitive data
